fix: retain floating conductances for integer shift arguments
An independent numeric-type regression exposes integer diagonal truncation
when the toy assembler receives shift=5 rather than 5.0. Allocate its diagonal
explicitly as float. All predeclared scientific shifts were already floats,
so no scientific matrix, source, tolerance or budget changes. Added regression
failed before this one-line fix; all 17 tests pass and the source tree matches
74411b5e1731f294bb4af1a2c9f717105229a26a. No production modifications.

+++ playground/defect_operator/operators.py
@@ -15,7 +15,7 @@ def assemble(k: np.ndarray, shift: float) -> sp.csr_matrix:
         raise ValueError('k must be square, with at least two cells per axis')
     if not np.all(np.isfinite(k)) or np.any(k <= 0) or not np.isfinite(shift) or shift < 0:
         raise ValueError('positive finite conductivities and nonnegative finite shift required')
-    n=k.shape[0]; indices=np.arange(n*n).reshape(n,n); diag=np.full(n*n,shift)
+    n=k.shape[0]; indices=np.arange(n*n).reshape(n,n); diag=np.full(n*n,shift,dtype=float)
     rows=[]; cols=[]; data=[]
     for axis in (0,1):
         s0=[slice(None),slice(None)]; s1=s0.copy()
